# OPS/openapi/300-ops-api.yaml
openapi: 3.1.0
info:
  title: 300 OPS Platform API
  description: |
    300 OPS is an engineering platform for smart buildings. This API provides
    comprehensive access to project management, engineering evaluation,
    prescription management, procurement, document generation, integrations,
    and AI services.
  version: 1.0.0
  contact:
    name: 300 OPS Support
    email: api@300ops.com
  license:
    name: Proprietary
    url: https://300ops.com/license
servers:
  - url: https://api.300ops.com/v1
    description: Production server
  - url: https://staging-api.300ops.com/v1
    description: Staging server
security:
  - BearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      description: Retrieve a paginated, filterable, sortable list of projects.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/FilterParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, on_hold, completed, archived]
        - name: client_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjects'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeProject'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
  /projects/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPathParam'
    get:
      tags: [Projects]
      summary: Get project by ID
      responses:
        '200':
          description: Project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeProject'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Projects]
      summary: Update project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeProject'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Projects]
      summary: Delete project
      responses:
        '204':
          description: Project deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /projects/{id}/buildings:
    parameters:
      - $ref: '#/components/parameters/UUIDPathParam'
    get:
      tags: [Projects, Buildings]
      summary: List buildings in a project
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: List of buildings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBuildings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /projects/{id}/requirements:
    parameters:
      - $ref: '#/components/parameters/UUIDPathParam'
    get:
      tags: [Projects, Requirements]
      summary: List requirements for a project
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of requirements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRequirements'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /projects/{id}/experiences:
    parameters:
      - $ref: '#/components/parameters/UUIDPathParam'
    get:
      tags: [Projects, Experiences]
      summary: List experiences for a project
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of experiences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExperiences'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /experiences:
    get:
      tags: [Experiences]
      summary: List experiences
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/FilterParam'
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of experiences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExperiences'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Experiences]
      summary: Create experience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperienceCreate'
      responses:
        '201':
          description: Experience created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeExperience'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
  /experiences/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPathParam'
    get:
      tags: [Experiences]
      summary: Get experience by ID
      responses:
        '200':
          description: Experience found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeExperience'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Experiences]
      summary: Update experience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperienceUpdate'
      responses:
        '200':
          description: Experience updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeExperience'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Experiences]
      summary: Delete experience
      responses:
        '204':
          description: Experience deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────────────────
  # SYSTEMS
  # ─────────────────────────────────────────────────────────
...
  /technical-zones:
    get:
      tags: [Technical Zones]
      summary: List technical zones
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/FilterParam'
        - name: building_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of technical zones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTechnicalZones'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Technical Zones]
      summary: Create technical zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicalZoneCreate'
      responses:
        '201':
          description: Technical zone created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeTechnicalZone'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
  /technical-zones/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPathParam'
    get:
      tags: [Technical Zones]
      summary: Get technical zone by ID
      responses:
        '200':
          description: Technical zone found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeTechnicalZone'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Technical Zones]
      summary: Update technical zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TechnicalZoneUpdate'
      responses:
        '200':
          description: Technical zone updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeTechnicalZone'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Technical Zones]
      summary: Delete technical zone
      responses:
        '204':
          description: Technical zone deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────────────────
  # PROCUREMENT / VALUE ENGINEERING
  # ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  parameters:
    UUIDPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique resource identifier
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        example: created_at:desc
      description: Sort field and direction (field:asc or field:desc)
    FilterParam:
      name: filter
      in: query
      schema:
        type: string
      description: JSON-encoded filter object
  schemas:
...
    BuildingType:
      type: string
      enum:
        - residential
        - commercial
        - industrial
        - mixed_use
        - healthcare
        - education
        - hospitality
        - data_center
        - retail
        - office
    SolutionLevel:
      type: string
      enum:
        - basic
        - standard
        - premium
        - custom
...
    Links:
      type: object
      required: [self]
      properties:
        self:
          type: string
          format: uri
        related:
          type: object
          additionalProperties:
            type: string
            format: uri
    Envelope:
      type: object
      required: [data, meta, links]
      properties:
        data:
          type: object
        meta:
          $ref: '#/components/schemas/Meta'
        links:
          $ref: '#/components/schemas/Links'

    # ───────────────────────────────────────────────
    # ERROR SCHEMAS
    # ───────────────────────────────────────────────
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        field:
          type: string
        details:
          type: object
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'

    # ───────────────────────────────────────────────
    # AUTH SCHEMAS
    # ───────────────────────────────────────────────
...
    Project:
      type: object
      required: [id, name, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        client_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, active, on_hold, completed, archived]
        building_type:
          $ref: '#/components/schemas/BuildingType'
        solution_level:
          $ref: '#/components/schemas/SolutionLevel'
        location:
          type: string
        start_date:
          type: string
          format: date
        target_end_date:
          type: string
          format: date
        budget_estimate:
          type: number
        currency:
          type: string
          default: EUR
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
    ProjectCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        client_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, active, on_hold, completed, archived]
          default: draft
        building_type:
          $ref: '#/components/schemas/BuildingType'
        solution_level:
          $ref: '#/components/schemas/SolutionLevel'
        location:
          type: string
        start_date:
          type: string
          format: date
        target_end_date:
          type: string
          format: date
        budget_estimate:
          type: number
        currency:
          type: string
          default: EUR
        metadata:
          type: object
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        client_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [draft, active, on_hold, completed, archived]
        building_type:
          $ref: '#/components/schemas/BuildingType'
        solution_level:
          $ref: '#/components/schemas/SolutionLevel'
        location:
          type: string
        start_date:
          type: string
          format: date
        target_end_date:
          type: string
          format: date
        budget_estimate:
          type: number
        currency:
          type: string
        metadata:
          type: object
    EnvelopeProject:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Project'
    PaginatedProjects:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Project'

    # ───────────────────────────────────────────────
    # BUILDING SCHEMAS
    # ───────────────────────────────────────────────
    Building:
      type: object
      required: [id, project_id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        building_type:
          $ref: '#/components/schemas/BuildingType'
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        total_floors:
          type: integer
        total_area_sqm:
          type: number
        year_built:
          type: integer
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PaginatedBuildings:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Building'

    # ───────────────────────────────────────────────
    # FLOOR SCHEMAS
    # ───────────────────────────────────────────────
...
    Requirement:
      type: object
      required: [id, project_id, title, category, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/RequirementCategory'
        priority:
          type: string
          enum: [low, medium, high, critical]
        detection_state:
          $ref: '#/components/schemas/DetectionState'
        detected_by:
          type: string
          enum: [user, ai, import, rule_engine]
        source_text:
          type: string
        related_system_id:
          type: string
          format: uuid
        room_ids:
          type: array
          items:
            type: string
            format: uuid
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PaginatedRequirements:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Requirement'

    # ───────────────────────────────────────────────
    # EXPERIENCE SCHEMAS
    # ───────────────────────────────────────────────
    Experience:
      type: object
      required: [id, project_id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        experience_type:
          type: string
          enum: [comfort, productivity, wellness, safety, entertainment, sustainability]
        target_metrics:
          type: object
        room_ids:
          type: array
          items:
            type: string
            format: uuid
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExperienceCreate:
      type: object
      required: [project_id, name]
      properties:
        project_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        experience_type:
          type: string
          enum: [comfort, productivity, wellness, safety, entertainment, sustainability]
        target_metrics:
          type: object
        room_ids:
          type: array
          items:
            type: string
            format: uuid
        metadata:
          type: object
    ExperienceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        experience_type:
          type: string
          enum: [comfort, productivity, wellness, safety, entertainment, sustainability]
        target_metrics:
          type: object
        room_ids:
          type: array
          items:
            type: string
            format: uuid
        metadata:
          type: object
    EnvelopeExperience:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Experience'
    PaginatedExperiences:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Experience'

    # ───────────────────────────────────────────────
    # SYSTEM SCHEMAS
    # ───────────────────────────────────────────────
...
    TechnicalZoneCreate:
      type: object
      required: [project_id, name]
      properties:
        project_id:
          type: string
          format: uuid
        building_id:
          type: string
          format: uuid
        name:
          type: string
        zone_type:
          type: string
        description:
          type: string
        floor_ids:
          type: array
          items:
            type: string
            format: uuid
        room_ids:
          type: array
          items:
            type: string
            format: uuid
        systems:
          type: array
          items:
            type: string
        metadata:
          type: object
    TechnicalZoneUpdate:
      type: object
      properties:
        building_id:
          type: string
          format: uuid
        name:
          type: string
        zone_type:
          type: string
        description:
          type: string
        floor_ids:
          type: array
          items:
            type: string
            format: uuid
        room_ids:
          type: array
          items:
            type: string
            format: uuid
        systems:
          type: array
          items:
            type: string
        metadata:
          type: object
    EnvelopeTechnicalZone:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TechnicalZone'
...
    PaginatedTechnicalZones:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/TechnicalZone'
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unprocessable:
      description: Unprocessable Entity - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'